Disabled Debug Mode for Flask Backend
this same change must be done to the GitHub Repo used to host the Prognosis Models in Heroku. The Issue was pointed out by CodeQL with the message "A Flask app appears to be run in debug mode. This may allow an attacker to run arbitrary code through the debugger."